Authentification correction

+++ back_end/back_gestion_immobiliere/src/main/java/com/intiformation/gestion/immo/dao/UserDAO.java
@@ -1,10 +1,13 @@
 package com.intiformation.gestion.immo.dao;
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.intiformation.gestion.immo.modele.UserAdmin;
 
+
 @Repository
-public interface UserDAO extends JpaRepository<UserAdmin, Integer> {
+public interface UserDAO extends JpaRepository<UserAdmin, Long> {
 
 }
